ser/mysite/requirements.txt
Django
pymongo == 3.6
six
mongoengine
djongo
djangorestframework == 3.8.2
django-filter 
uWSGI 